Add support for a preloader to the Flash build process.
Matrix had to be changed because of the issue already reported in #143 and #155, or else the compiled preloaded swf will not run without "Error #1014: Class ArrayAccess could not be found." ArrayAccess appears to be optimized out...or is otherwise unavailable. I'm not sure it's being used correctly in Matrix.transformArray anyway, and my change doesn't appear to hurt.