Move everything that I can, in the most inflexible way, to 30 fps from 60 fps.
In any of the testing that I've done, 60 fps is inconsistent in the browser, especially if there are other Flash instances running in the background. This causes frequent and noticable framerate drops that are especially visible with fast and smoothly moving objects. I would like to make framerate configurable via flambe.yaml, but it would require a macro to get that value over to the relevant Flambe sources. Such an implementation might call for a more general approach to getting configurable values from flambe.yaml into the code, like what is shown here: http://blog.onthewings.net/2012/09/02/haxe-advanced-conditional-compilation/.